LayoutTests/platform/chromium/fast/js/cyclic-prototypes-expected.txt

CONSOLE MESSAGE: line 16: Uncaught Error: Cyclic __proto__ value
This test makes sure we don't hang when setting cyclic prototype values: http://bugs.webkit.org/show_bug.cgi?id=6985

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS o1.__proto__ = o3; threw exception Error: Cyclic __proto__ value.
PASS ({}).hasOwnProperty.call(o1, '__proto__') is false
FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
